Refactor BE code
