+++ app/Providers/RepositoryServiceProvider.php
@@ -2,10 +2,17 @@
 
 namespace App\Providers;
 
-use App\Interfaces\DocumentRepositoryInterface;
+use App\Interfaces\Repository\DocumentRepositoryInterface;
+use App\Interfaces\Repository\TemplateRepositoryInterface;
+
+use App\Interfaces\Service\DocumentServiceInterface;
+use App\Interfaces\Service\TemplateServiceInterface;
+
 use App\Repositories\TemplateRepository;
-use App\Interfaces\TemplateRepositoryInterface;
 use App\Repositories\DocumentRepository;
+
+use App\Services\DocumentService;
+use App\Services\TemplateService;
 use Illuminate\Support\ServiceProvider;
 
 class RepositoryServiceProvider extends ServiceProvider
@@ -19,6 +26,9 @@ class RepositoryServiceProvider extends ServiceProvider
     {
         $this->app->bind(TemplateRepositoryInterface::class, TemplateRepository::class);
         $this->app->bind(DocumentRepositoryInterface::class, DocumentRepository::class);
+
+        $this->app->bind(DocumentServiceInterface::class, DocumentService::class);
+        $this->app->bind(TemplateServiceInterface::class, TemplateService::class);
     }
 
     /**
